edited BWEB pathfinding for new distance func; a few edits in Movement::move_group()

+++ FrogFish/Source/movement/Move.h
@@ -55,7 +55,7 @@ namespace Movement::Move {
                 BWAPI::TilePosition dest,
                 bool attack=false,
                 bool wait=false,
-                float group_cohesion=C_MED
+                double group_cohesion=C_MED
             );
 
     // Turns a single unit into a group of one and calls the std::vector<BWAPI::Unit> move()
